pyproject.toml
[project]
name = "ai-exp"
version = "0.1.0"
description = "Experience-based memory system with affect-aware retrieval"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "sqlmodel>=0.0.22",
    "chromadb>=0.5.23",
    "sentence-transformers>=3.3.1",
    "numpy>=2.2.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.4",
    "black>=24.10.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.4",
    "black>=24.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
